+++ src/lib.rs
@@ -108,143 +108,331 @@
 #![cfg_attr(not(test), no_std)]
 #![deny(missing_debug_implementations, missing_docs)]
 
+#[cfg(feature = "alloc")]
 extern crate alloc;
 
 use crossbeam_utils::CachePadded;
 
+#[cfg(feature = "alloc")]
 use alloc::sync::Arc;
 use core::{
     cell::UnsafeCell,
-    fmt,
+    fmt::{self, Debug, Display, Formatter},
     ops::{Deref, DerefMut},
-    sync::atomic::{AtomicU8, Ordering},
+    sync::atomic::{AtomicU8, AtomicUsize, Ordering},
 };
 
 /// A triple buffer, useful for nonblocking and thread-safe data sharing
 ///
 /// A triple buffer is a single-producer single-consumer nonblocking
 /// communication channel which behaves like a shared variable: the producer
-/// submits regular updates, and the consumer accesses the latest available
-/// value whenever it feels like it.
+/// thread can submit regular updates, and the consumer thread can access the
+/// latest available value whenever it feels like it.
 #[derive(Debug)]
+#[repr(C)]
 pub struct TripleBuffer<T: Send> {
-    /// Input object used by producers to send updates
-    input: Input<T>,
+    /// Data storage buffers
+    ///
+    /// Should be the first member to allow for maximal `#[link_section]`
+    /// embedded DMA shenanigans. This is enforces via `repr(C)`.
+    buffers: [CachePadded<UnsafeCell<T>>; 3],
 
-    /// Output object used by consumers to read the current value
-    output: Output<T>,
+    /// Atomic state used to synchronize access to buffers
+    status: CachePadded<Status>,
 }
 //
 impl<T: Clone + Send> TripleBuffer<T> {
-    /// Construct a triple buffer with a certain initial value
+    /// Set up a triple buffer with a certain initial value
     pub fn new(initial: &T) -> Self {
-        Self::new_impl(|| initial.clone())
+        Self::from_fn(|| initial.clone())
     }
 }
 //
 impl<T: Default + Send> Default for TripleBuffer<T> {
-    /// Construct a triple buffer with a default-constructed value
+    /// Set up a triple buffer with a default-constructed initial value
     fn default() -> Self {
-        Self::new_impl(T::default)
+        Self::from_fn(T::default)
     }
 }
 //
 impl<T: Send> TripleBuffer<T> {
-    /// Construct a triple buffer, using a functor to generate initial values
-    fn new_impl(mut generator: impl FnMut() -> T) -> Self {
-        // Start with the shared state...
-        let shared_state = Arc::new(SharedState::new(|_i| generator(), 0));
+    /// Set up a triple buffer, generating the three inner initial values
+    /// using the `generator` function
+    ///
+    /// The values that are produced by the `generator` functions should be
+    /// _logically identical_, in the sense that...
+    ///
+    /// - If `T` implements `PartialEq`, then these values should compare equal
+    ///   to each other.
+    /// - You should not need to know which of these values the triple buffer's
+    ///   producer and consumer interfaces will end up initially pointing to.
+    ///   For most purpose, you should consider them to be the same value.
+    ///
+    /// It is, however, acceptable for these values to be _bitwise distinct_,
+    /// for example they can be pointers to three distinct allocations that have
+    /// identical content.
+    pub fn from_fn(mut generator: impl FnMut() -> T) -> Self {
+        Self::from_values(core::array::from_fn(|_| generator()))
+    }
 
-        // ...then construct the input and output structs
-        TripleBuffer {
-            input: Input {
-                shared: shared_state.clone(),
-                input_idx: 1,
-            },
-            output: Output {
-                shared: shared_state,
-                output_idx: 2,
-            },
+    /// Wrap three user-provided values into a `TripleBuffer`
+    ///
+    /// As with [`from_fn()`](Self::from_fn), the three input values should be
+    /// _logically identical_, but can be _bitwise distinct_.
+    pub fn from_values(values: [T; 3]) -> Self {
+        Self {
+            buffers: values.map(|v| CachePadded::new(UnsafeCell::new(v))),
+            status: CachePadded::new(Status::new()),
         }
     }
 
-    /// Extract input and output of the triple buffer
-    //
-    // NOTE: Although it would be nicer to directly return `Input` and `Output`
-    //       from `new()`, the `split()` design gives some API evolution
-    //       headroom towards future allocation-free modes of operation where
-    //       the SharedState is a static variable, or a stack-allocated variable
-    //       used through scoped threads or other unsafe thread synchronization.
-    //
-    //       See https://github.com/HadrienG2/triple-buffer/issues/8 .
-    //
-    pub fn split(self) -> (Input<T>, Output<T>) {
-        (self.input, self.output)
+    /// Set up the [`Input`] and [`Output`] of a reference-counted triple buffer
+    ///
+    /// This method is a common-case convenience shortcut for the
+    /// [`inout()`](TripleBufferRef::inout) method of the [`TripleBufferRef`]
+    /// trait, which would otherwise require you to import said trait and use
+    /// the awkward fully qualified function call syntax for clarity.
+    ///
+    /// # Panics
+    ///
+    /// Panics if this triple buffer already has an [`Input`] or [`Output`]
+    #[cfg(feature = "alloc")]
+    pub fn inout_from_arc(self: Arc<Self>) -> (Input<Arc<Self>>, Output<Arc<Self>>) {
+        <Arc<Self> as TripleBufferRef>::inout(self)
+    }
+
+    /// Set up the [`Input`] and [`Output`] of a triple buffer from a shared
+    /// reference
+    ///
+    /// This method is a common-case convenience shortcut for the
+    /// [`inout()`](TripleBufferRef::inout) method of the [`TripleBufferRef`]
+    /// trait, which would otherwise require you to import said trait and use
+    /// the awkward fully qualified function call syntax for clarity.
+    ///
+    /// Beware that when the producer and consumer interfaces are set up from a
+    /// shared reference like this, they cannot be used outside of the scope
+    /// where the underlying triple buffer is defined. In naive usage, this will
+    /// often be too limiting for inter-thread communication.
+    ///
+    /// However, this set up method works for `no_alloc` use cases where
+    /// synchronization primitives must commonly be allocated as static
+    /// variables due to lack of heap allocations. In this case, the `'static`
+    /// lifetime of the associated references will let you use the resulting
+    /// `Input`s and `Output`s anywhere in the program.
+    ///
+    /// Another situation where this method can be used is "scoped thread"
+    /// abstractions like `std::thread::scope()`. To support these, `'a` is
+    /// not constrained to be `'static` even though that will often be what you
+    /// want when spawning threads in a more traditional way.
+    ///
+    /// # Panics
+    ///
+    /// Panics if this triple buffer already has an [`Input`] or [`Output`]
+    pub fn inout_from_ref<'a>(&'a self) -> (Input<&'a Self>, Output<&'a Self>) {
+        <&'a Self as TripleBufferRef>::inout(self)
     }
 }
 //
-/// Shorthand for `TripleBuffer::new(initial).split()`
-pub fn triple_buffer<T: Clone + Send>(initial: &T) -> (Input<T>, Output<T>) {
-    TripleBuffer::new(initial).split()
+/// Convenience shortcut to set up a reference-counted triple buffer and get
+/// the associated producer and consumer interfaces
+#[cfg(feature = "alloc")]
+pub fn triple_buffer<T: Clone + Send>(
+    initial: &T,
+) -> (Input<Arc<TripleBuffer<T>>>, Output<Arc<TripleBuffer<T>>>) {
+    Arc::new(TripleBuffer::new(initial)).inout_from_arc()
 }
 //
-// The Clone and PartialEq traits are used internally for testing and I don't
+impl<T: Send> Drop for TripleBuffer<T> {
+    fn drop(&mut self) {
+        let next_io_indices = self.status.next_io_indices.load(Ordering::Relaxed);
+        let next_input_idx =
+            (next_io_indices & NEXT_INPUT_MASK) >> NEXT_INPUT_MASK.trailing_zeros();
+        let next_output_idx =
+            (next_io_indices & NEXT_OUTPUT_MASK) >> NEXT_OUTPUT_MASK.trailing_zeros();
+        for next_idx in [next_input_idx, next_output_idx] {
+            debug_assert_ne!(
+                next_idx, NO_BUFFER_IDX,
+                "No Input or Output should exist at TripleBuffer destruction time"
+            );
+        }
+    }
+}
+//
+unsafe impl<T: Send> Sync for TripleBuffer<T> {}
+//
+// Cloning and equality are used internally for testing and I don't
 // want to commit to supporting them publicly for now.
-//
 #[doc(hidden)]
-impl<T: Clone + Send> Clone for TripleBuffer<T> {
-    fn clone(&self) -> Self {
-        // Clone the shared state. This is safe because at this layer of the
-        // interface, one needs an Input/Output &mut to mutate the shared state.
-        let shared_state = Arc::new(unsafe { (*self.input.shared).clone() });
-
-        // ...then the input and output structs
-        TripleBuffer {
-            input: Input {
-                shared: shared_state.clone(),
-                input_idx: self.input.input_idx,
-            },
-            output: Output {
-                shared: shared_state,
-                output_idx: self.output.output_idx,
-            },
+impl<T: Clone + Send> TripleBuffer<T> {
+    /// Like `Clone`, but requires exclusive access for safety
+    fn seq_clone(&mut self) -> Self {
+        Self {
+            buffers: core::array::from_fn(|i| {
+                CachePadded::new(UnsafeCell::new(self.buffers[i].get_mut().clone()))
+            }),
+            status: self.status.clone(),
         }
     }
 }
 //
 #[doc(hidden)]
-impl<T: PartialEq + Send> PartialEq for TripleBuffer<T> {
-    fn eq(&self, other: &Self) -> bool {
-        // Compare the shared states. This is safe because at this layer of the
-        // interface, one needs an Input/Output &mut to mutate the shared state.
-        let shared_states_equal = unsafe { (*self.input.shared).eq(&*other.input.shared) };
+impl<T: PartialEq + Send> TripleBuffer<T> {
+    /// Like `Eq`, but requires exclusive access for safety.
+    fn seq_eq(&mut self, other: &mut Self) -> bool {
+        // Check whether the contents of all buffers are equal...
+        let buffers_equal = self
+            .buffers
+            .iter_mut()
+            .zip(&mut other.buffers)
+            .all(|(cell1, cell2)| -> bool { cell1.get_mut() == cell2.get_mut() });
 
-        // Compare the rest of the triple buffer states
-        shared_states_equal
-            && (self.input.input_idx == other.input.input_idx)
-            && (self.output.output_idx == other.output.output_idx)
+        // ...then check whether the rest of the shared state is equal
+        buffers_equal && self.status == other.status
+    }
+}
+
+/// Access to a [`TripleBuffer`] from a shared reference or [`Arc`] thereof
+///
+/// While it may seem like this trait is unnecessary, it is actually needed to
+/// keep the generic types of `Input` and `Output` simple and to allow for a
+/// generic implementation of `inout()`.
+pub trait TripleBufferRef: Clone + Deref + Send {
+    /// Type of data that is being shared via this [`TripleBuffer`], i.e. the
+    /// `T` argument to `TripleBuffer<T>`.
+    type T: Send;
+
+    /// Set up the [`Input`] and [`Output`] of this triple buffer in a single
+    /// transaction
+    ///
+    /// This is more efficient than calling [`Input::new()`] and
+    /// [`Output::new()`] separately. For maximal efficiency, you should also
+    /// keep `Input`s and `Output`s alive as long as possible, ideally for the
+    /// entire lifetime of the underlying triple buffer.
+    ///
+    /// Because this trait is implemented for multiple reference types and
+    /// Rust's auto-deref makes calling trait methods on reference types
+    /// confusing, it is strongly advised to call this method using the `<Ref as
+    /// TripleBufferRef>::inout(...)` fully qualified syntax.
+    ///
+    /// # Panics
+    ///
+    /// Panics if this triple buffer already has an [`Input`] or [`Output`]
+    fn inout(self) -> (Input<Self>, Output<Self>) {
+        // Atomically fetch and clear the next input and output buffer indices
+        let io_indices = self
+            .status()
+            .next_io_indices
+            .swap(NEXT_NOTHING, Ordering::Acquire);
+
+        // Decode input and output indices and make sure that no Input and
+        // Output currently exists for this triple buffer
+        let input_idx = (io_indices & NEXT_INPUT_MASK) >> NEXT_INPUT_MASK.trailing_zeros();
+        let output_idx = (io_indices & NEXT_OUTPUT_MASK) >> NEXT_OUTPUT_MASK.trailing_zeros();
+        for io_idx in [input_idx, output_idx] {
+            assert_ne!(
+                io_idx, NO_BUFFER_IDX,
+                "This triple buffer already has an Input or Output"
+            );
+        }
+
+        // Set up the producer and consumer interfaces
+        let input = Input {
+            shared: self.clone(),
+            input_idx,
+        };
+        (
+            input,
+            Output {
+                shared: self,
+                output_idx,
+            },
+        )
+    }
+
+    /// Internal access to the private `buffers` member
+    #[doc(hidden)]
+    fn buffers(&self) -> &[CachePadded<UnsafeCell<Self::T>>; 3];
+
+    /// Internal access to the private `status` member
+    #[allow(private_interfaces)]
+    #[doc(hidden)]
+    fn status<'a>(&'a self) -> &'a Status
+    where
+        Self::T: 'a;
+}
+//
+impl<T, Ref> TripleBufferRef for Ref
+where
+    Ref: Clone + Deref<Target = TripleBuffer<T>> + Send,
+    T: Send,
+{
+    type T = T;
+
+    #[inline(always)]
+    fn buffers(&self) -> &[CachePadded<UnsafeCell<T>>; 3] {
+        &self.buffers
+    }
+
+    #[allow(private_interfaces)]
+    #[inline(always)]
+    fn status<'a>(&'a self) -> &'a Status
+    where
+        Self::T: 'a,
+    {
+        &self.status
     }
 }
 
 /// Producer interface to the triple buffer
 ///
-/// The producer of data can use this struct to submit updates to the triple
+/// The producer thread can use this struct to submit updates to the triple
 /// buffer whenever he likes. These updates are nonblocking: a collision between
 /// the producer and the consumer will result in cache contention, but deadlocks
 /// and scheduling-induced slowdowns cannot happen.
 #[derive(Debug)]
-pub struct Input<T: Send> {
-    /// Reference-counted shared state
-    shared: Arc<SharedState<T>>,
+pub struct Input<Shared: TripleBufferRef> {
+    /// Reference to the shared state
+    shared: Shared,
 
     /// Index of the input buffer (which is private to the producer)
-    input_idx: BufferIndex,
+    input_idx: BufferIdx,
 }
 //
 // Public interface
-impl<T: Send> Input<T> {
+impl<Shared: TripleBufferRef> Input<Shared> {
+    /// Set up the producer interface of a triple buffer
+    ///
+    /// It is more efficient to set up both the producer and consumer interfaces
+    /// at the same time using [`TripleBufferRef::inout()`] and its convenience
+    /// shortcuts [`TripleBuffer::inout_from_arc()`] and
+    /// [`TripleBuffer::inout_from_ref()`].
+    ///
+    /// # Panics
+    ///
+    /// Panics if this triple buffer already has an [`Input`]
+    pub fn new(shared: Shared) -> Self {
+        let mut input_idx = NO_BUFFER_IDX;
+        shared
+            .status()
+            .next_io_indices
+            .fetch_update(Ordering::Release, Ordering::Relaxed, |old_next_indices| {
+                input_idx =
+                    (old_next_indices & NEXT_INPUT_MASK) >> NEXT_INPUT_MASK.trailing_zeros();
+                assert_ne!(
+                    input_idx, NO_BUFFER_IDX,
+                    "This triple buffer already has an Input"
+                );
+                Some(
+                    (old_next_indices & !NEXT_INPUT_MASK)
+                        | (NO_BUFFER_IDX << NEXT_INPUT_MASK.trailing_zeros()),
+                )
+            })
+            .expect("Not allowed to fail");
+        Self { shared, input_idx }
+    }
+
     /// Write a new value into the triple buffer
-    pub fn write(&mut self, value: T) {
+    pub fn write(&mut self, value: Shared::T) {
         // Update the input buffer
         *self.input_buffer_mut() = value;
 
@@ -260,8 +448,8 @@ impl<T: Send> Input<T> {
     /// implementation. If what you truly need is a double buffer, build
     /// yourself a proper blocking one instead of wasting CPU time.
     pub fn consumed(&self) -> bool {
-        let back_info = self.shared.back_info.load(Ordering::Relaxed);
-        back_info & BACK_DIRTY_BIT == 0
+        let back_info = self.shared.status().back_info.load(Ordering::Relaxed);
+        back_info & BACK_UPDATED_BIT == 0
     }
 
     /// Query the current value of the input buffer
@@ -270,10 +458,10 @@ impl<T: Send> Input<T> {
     /// [`input_buffer_mut()`](Input::input_buffer_mut). Please read the
     /// documentation of that method for more information on the precautions
     /// that need to be taken when accessing the input buffer in place.
-    pub fn input_buffer(&self) -> &T {
+    pub fn input_buffer(&self) -> &Shared::T {
         // This is safe because the synchronization protocol ensures that we
         // have exclusive access to this buffer.
-        let input_ptr = self.shared.buffers[self.input_idx as usize].get();
+        let input_ptr = self.shared.buffers()[self.input_idx as usize].get();
         unsafe { &*input_ptr }
     }
 
@@ -300,10 +488,10 @@ impl<T: Send> Input<T> {
     /// input buffer reference goes out of scope, consider using the
     /// [`input_buffer_publisher()`](Input::input_buffer_publisher) RAII
     /// interface instead.
-    pub fn input_buffer_mut(&mut self) -> &mut T {
+    pub fn input_buffer_mut(&mut self) -> &mut Shared::T {
         // This is safe because the synchronization protocol ensures that we
         // have exclusive access to this buffer.
-        let input_ptr = self.shared.buffers[self.input_idx as usize].get();
+        let input_ptr = self.shared.buffers()[self.input_idx as usize].get();
         unsafe { &mut *input_ptr }
     }
 
@@ -335,17 +523,16 @@ impl<T: Send> Input<T> {
         //     before that index is actually read, as well as on weird hardware
         //     with incoherent caches like GPUs or old DEC Alpha where keeping
         //     data in sync across cores requires manual action.
-        //
-        let former_back_info = self
-            .shared
-            .back_info
-            .swap(self.input_idx | BACK_DIRTY_BIT, Ordering::AcqRel);
+        let former_back_info = self.shared.status().back_info.swap(
+            self.input_idx as BackInfo | BACK_UPDATED_BIT,
+            Ordering::AcqRel,
+        );
 
         // The old back buffer becomes our new input buffer
-        self.input_idx = former_back_info & BACK_INDEX_MASK;
+        self.input_idx = (former_back_info & BACK_IDX_MASK) as BufferIdx;
 
         // Tell whether we have overwritten unread data
-        former_back_info & BACK_DIRTY_BIT != 0
+        former_back_info & BACK_UPDATED_BIT != 0
     }
 
     /// Access the input buffer wrapped in the `InputPublishGuard`
@@ -360,8 +547,31 @@ impl<T: Send> Input<T> {
     ///
     /// [`input_buffer_mut()`]: Input::input_buffer_mut
     /// [`publish()`]: Input::publish
-    pub fn input_buffer_publisher(&mut self) -> InputPublishGuard<T> {
-        InputPublishGuard { reference: self }
+    pub fn input_buffer_publisher(&mut self) -> InputPublishGuard<Shared> {
+        InputPublishGuard(self)
+    }
+}
+//
+impl<Shared: TripleBufferRef> Drop for Input<Shared> {
+    fn drop(&mut self) {
+        // Save the current input buffer index for the next time an `Input`
+        // will be built and check state consistency in debug builds.
+        self.shared
+            .status()
+            .next_io_indices
+            .fetch_update(Ordering::Release, Ordering::Relaxed, |old_next_indices| {
+                let old_next_input_idx =
+                    (old_next_indices & NEXT_INPUT_MASK) >> NEXT_INPUT_MASK.trailing_zeros();
+                debug_assert_eq!(
+                    old_next_input_idx, NO_BUFFER_IDX,
+                    "The shared state should not contain a next input index while an Input is live"
+                );
+                Some(
+                    (old_next_indices & !NEXT_INPUT_MASK)
+                        | (self.input_idx << NEXT_INPUT_MASK.trailing_zeros()),
+                )
+            })
+            .expect("Not allowed to fail");
     }
 }
 
@@ -374,62 +584,97 @@ impl<T: Send> Input<T> {
 /// automatically.
 ///
 /// This structure is created by the [`Input::input_buffer_publisher()`] method.
-pub struct InputPublishGuard<'a, T: 'a + Send> {
-    reference: &'a mut Input<T>,
-}
-
-impl<T: Send> Deref for InputPublishGuard<'_, T> {
-    type Target = T;
-
-    fn deref(&self) -> &T {
-        self.reference.input_buffer()
+pub struct InputPublishGuard<'a, Shared: TripleBufferRef>(&'a mut Input<Shared>);
+//
+impl<Shared: TripleBufferRef> Debug for InputPublishGuard<'_, Shared>
+where
+    Shared::T: Debug,
+{
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        Debug::fmt(&self.0.input_buffer(), f)
     }
 }
+//q
+impl<Shared: TripleBufferRef> Deref for InputPublishGuard<'_, Shared> {
+    type Target = Shared::T;
 
-impl<T: Send> DerefMut for InputPublishGuard<'_, T> {
-    fn deref_mut(&mut self) -> &mut T {
-        self.reference.input_buffer_mut()
+    fn deref(&self) -> &Shared::T {
+        self.0.input_buffer()
     }
 }
-
-impl<T: Send> Drop for InputPublishGuard<'_, T> {
+//
+impl<Shared: TripleBufferRef> DerefMut for InputPublishGuard<'_, Shared> {
+    fn deref_mut(&mut self) -> &mut Shared::T {
+        self.0.input_buffer_mut()
+    }
+}
+//
+impl<Shared: TripleBufferRef> Display for InputPublishGuard<'_, Shared>
+where
+    Shared::T: Display,
+{
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        Display::fmt(self.0.input_buffer(), f)
+    }
+}
+//
+impl<Shared: TripleBufferRef> Drop for InputPublishGuard<'_, Shared> {
     #[inline]
     fn drop(&mut self) {
-        self.reference.publish();
-    }
-}
-
-impl<T: Send + fmt::Debug> fmt::Debug for InputPublishGuard<'_, T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        fmt::Debug::fmt(&**self, f)
-    }
-}
-
-impl<T: fmt::Display + Send> fmt::Display for InputPublishGuard<'_, T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        (**self).fmt(f)
+        self.0.publish();
     }
 }
 
 /// Consumer interface to the triple buffer
 ///
-/// The consumer of data can use this struct to access the latest published
+/// The consumer thread can use this struct to access the latest published
 /// update from the producer whenever he likes. Readout is nonblocking: a
 /// collision between the producer and consumer will result in cache contention,
 /// but deadlocks and scheduling-induced slowdowns cannot happen.
 #[derive(Debug)]
-pub struct Output<T: Send> {
-    /// Reference-counted shared state
-    shared: Arc<SharedState<T>>,
+pub struct Output<Shared: TripleBufferRef> {
+    /// Reference to the shared state
+    shared: Shared,
 
     /// Index of the output buffer (which is private to the consumer)
-    output_idx: BufferIndex,
+    output_idx: BufferIdx,
 }
 //
 // Public interface
-impl<T: Send> Output<T> {
+impl<Shared: TripleBufferRef> Output<Shared> {
+    /// Set up the consumer interface of a triple buffer
+    ///
+    /// It is more efficient to set up both the producer and consumer interfaces
+    /// at the same time using [`TripleBufferRef::inout()`] and its convenience
+    /// shortcuts [`TripleBuffer::inout_from_arc()`] and
+    /// [`TripleBuffer::inout_from_ref()`].
+    ///
+    /// # Panics
+    ///
+    /// Panics if this triple buffer already has an [`Output`]
+    pub fn new(shared: Shared) -> Self {
+        let mut output_idx = NO_BUFFER_IDX;
+        shared
+            .status()
+            .next_io_indices
+            .fetch_update(Ordering::Release, Ordering::Relaxed, |old_next_indices| {
+                output_idx =
+                    (old_next_indices & NEXT_OUTPUT_MASK) >> NEXT_OUTPUT_MASK.trailing_zeros();
+                assert_ne!(
+                    output_idx, NO_BUFFER_IDX,
+                    "This triple buffer already has an Output"
+                );
+                Some(
+                    (old_next_indices & !NEXT_OUTPUT_MASK)
+                        | (NO_BUFFER_IDX << NEXT_OUTPUT_MASK.trailing_zeros()),
+                )
+            })
+            .expect("Not allowed to fail");
+        Self { shared, output_idx }
+    }
+
     /// Access the latest value from the triple buffer
-    pub fn read(&mut self) -> &T {
+    pub fn read(&mut self) -> &Shared::T {
         // Fetch updates from the producer
         self.update();
 
@@ -445,8 +690,8 @@ impl<T: Send> Output<T> {
     /// implementation. If what you truly need is a double buffer, build
     /// yourself a proper blocking one instead of wasting CPU time.
     pub fn updated(&self) -> bool {
-        let back_info = self.shared.back_info.load(Ordering::Relaxed);
-        back_info & BACK_DIRTY_BIT != 0
+        let back_info = self.shared.status().back_info.load(Ordering::Relaxed);
+        back_info & BACK_UPDATED_BIT != 0
     }
 
     /// Query the current value of the output buffer
@@ -456,7 +701,7 @@ impl<T: Send> Output<T> {
     /// instead, as `peek_output_buffer()` is scheduled for removal in the next
     /// major release of `triple-buffer`.
     #[deprecated = "Please use output_buffer() instead"]
-    pub fn peek_output_buffer(&self) -> &T {
+    pub fn peek_output_buffer(&self) -> &Shared::T {
         self.output_buffer()
     }
 
@@ -470,10 +715,10 @@ impl<T: Send> Output<T> {
     /// In particular, remember that this method does not update the output
     /// buffer automatically. You need to call [`update()`](Output::update) in
     /// order to fetch buffer updates from the producer.
-    pub fn output_buffer(&self) -> &T {
+    pub fn output_buffer(&self) -> &Shared::T {
         // This is safe because the synchronization protocol ensures that we
         // have exclusive access to this buffer.
-        let output_ptr = self.shared.buffers[self.output_idx as usize].get();
+        let output_ptr = self.shared.buffers()[self.output_idx as usize].get();
         unsafe { &*output_ptr }
     }
 
@@ -497,10 +742,10 @@ impl<T: Send> Output<T> {
     ///
     /// [`read()`]: Output::read
     /// [`update()`]: Output::update
-    pub fn output_buffer_mut(&mut self) -> &mut T {
+    pub fn output_buffer_mut(&mut self) -> &mut Shared::T {
         // This is safe because the synchronization protocol ensures that we
         // have exclusive access to this buffer.
-        let output_ptr = self.shared.buffers[self.output_idx as usize].get();
+        let output_ptr = self.shared.buffers()[self.output_idx as usize].get();
         unsafe { &mut *output_ptr }
     }
 
@@ -517,9 +762,6 @@ impl<T: Send> Output<T> {
         // Check if an update is present in the back-buffer
         let updated = self.updated();
         if updated {
-            // Access the shared state
-            let shared_state = &(*self.shared);
-
             // If so, exchange our output buffer with the back-buffer, thusly
             // acquiring exclusive access to the old back buffer while giving
             // the producer a new back-buffer to write to.
@@ -538,104 +780,186 @@ impl<T: Send> Output<T> {
             //     compiler or CPU correctly speculated the new buffer's index
             //     before that index is actually read, as well as on weird hardware
             //     like GPUs where CPU caches require manual synchronization.
-            //
-            let former_back_info = shared_state
+            let former_back_info = self
+                .shared
+                .status()
                 .back_info
-                .swap(self.output_idx, Ordering::AcqRel);
+                .swap(self.output_idx as BackInfo, Ordering::AcqRel);
 
             // Make the old back-buffer our new output buffer
-            self.output_idx = former_back_info & BACK_INDEX_MASK;
+            self.output_idx = (former_back_info & BACK_IDX_MASK) as BufferIdx;
         }
 
         // Tell whether an update was carried out
         updated
     }
 }
+//
+impl<Shared: TripleBufferRef> Drop for Output<Shared> {
+    fn drop(&mut self) {
+        // Save the current output buffer index for the next time an `Output`
+        // will be built and check state consistency in debug builds.
+        self.shared
+            .status()
+            .next_io_indices
+            .fetch_update(Ordering::Release, Ordering::Relaxed, |old_next_indices| {
+                let old_next_output_idx =
+                    (old_next_indices & NEXT_OUTPUT_MASK) >> NEXT_OUTPUT_MASK.trailing_zeros();
+                debug_assert_eq!(
+                    old_next_output_idx, NO_BUFFER_IDX,
+                    "The shared state should not contain a next output index while an Output is live"
+                );
+                Some(
+                    (old_next_indices & !NEXT_OUTPUT_MASK)
+                        | (self.output_idx << NEXT_OUTPUT_MASK.trailing_zeros()),
+                )
+            })
+            .expect("Not allowed to fail");
+    }
+}
 
-/// Triple buffer shared state
+/// Triple buffer status information
 ///
-/// In a triple buffering communication protocol, the producer and consumer
-/// share the following storage:
-///
-/// - Three memory buffers suitable for storing the data at hand
-/// - Information about the back-buffer: which buffer is the current back-buffer
-///   and whether an update was published since the last readout.
+/// Should be [`CachePadded`] to avoid false sharing with the data buffers.
 #[derive(Debug)]
-struct SharedState<T: Send> {
-    /// Data storage buffers
-    buffers: [CachePadded<UnsafeCell<T>>; 3],
+struct Status {
+    /// Back buffer index + truth that an update is pending
+    back_info: AtomicBackInfo,
 
-    /// Information about the current back-buffer state
-    back_info: CachePadded<AtomicBackBufferInfo>,
+    /// Next input and output indices, bit-packed in a single word
+    next_io_indices: AtomicNextIndices,
 }
 //
-#[doc(hidden)]
-impl<T: Send> SharedState<T> {
-    /// Given (a way to generate) buffer contents and the back info, build the shared state
-    fn new(mut gen_buf_data: impl FnMut(usize) -> T, back_info: BackBufferInfo) -> Self {
-        let mut make_buf = |i| -> CachePadded<UnsafeCell<T>> {
-            CachePadded::new(UnsafeCell::new(gen_buf_data(i)))
-        };
+impl Status {
+    fn new() -> Self {
         Self {
-            buffers: [make_buf(0), make_buf(1), make_buf(2)],
-            back_info: CachePadded::new(AtomicBackBufferInfo::new(back_info)),
+            back_info: AtomicBackInfo::new(BACK_INFO_INIT),
+            next_io_indices: AtomicNextIndices::new(NEXT_INDICES_INIT),
         }
     }
 }
 //
-#[doc(hidden)]
-impl<T: Clone + Send> SharedState<T> {
-    /// Cloning the shared state is unsafe because you must ensure that no one
-    /// is concurrently accessing it, since &self is enough for writing.
-    unsafe fn clone(&self) -> Self {
-        Self::new(
-            |i| (*self.buffers[i].get()).clone(),
-            self.back_info.load(Ordering::Relaxed),
-        )
+impl Clone for Status {
+    /// Meant for sequential testing code paths only
+    ///
+    /// Use on a [`Status`] that is being modified concurrently will result in
+    /// unpredictable results, including observation of inconsistent states.
+    fn clone(&self) -> Self {
+        Self {
+            back_info: AtomicBackInfo::new(self.back_info.load(Ordering::Relaxed)),
+            next_io_indices: AtomicNextIndices::new(self.next_io_indices.load(Ordering::Relaxed)),
+        }
     }
 }
 //
-#[doc(hidden)]
-impl<T: PartialEq + Send> SharedState<T> {
-    /// Equality is unsafe for the same reason as cloning: you must ensure that
-    /// no one is concurrently accessing the triple buffer to avoid data races.
-    unsafe fn eq(&self, other: &Self) -> bool {
-        // Check whether the contents of all buffers are equal...
-        let buffers_equal = self
-            .buffers
-            .iter()
-            .zip(other.buffers.iter())
-            .all(|tuple| -> bool {
-                let (cell1, cell2) = tuple;
-                *cell1.get() == *cell2.get()
-            });
-
-        // ...then check whether the rest of the shared state is equal
-        buffers_equal
-            && (self.back_info.load(Ordering::Relaxed) == other.back_info.load(Ordering::Relaxed))
+impl Default for Status {
+    fn default() -> Self {
+        Self::new()
     }
 }
 //
-unsafe impl<T: Send> Sync for SharedState<T> {}
+impl PartialEq for Status {
+    /// Meant for sequential testing code paths only
+    ///
+    /// Use on a [`Status`] that is being modified concurrently will result in
+    /// unpredictable results, including observation of inconsistent states.
+    fn eq(&self, other: &Self) -> bool {
+        self.back_info.load(Ordering::Relaxed) == other.back_info.load(Ordering::Relaxed)
+            && self.next_io_indices.load(Ordering::Relaxed)
+                == other.next_io_indices.load(Ordering::Relaxed)
+    }
+}
 
-// Index types used for triple buffering
+/// Type used for buffer indices (in range 0..=2)
+type BufferIdx = u8;
+
+/// Bitfield that tracks...
+///
+/// - The index of the current back buffer, i.e. the buffer which neither the
+///   producer nor the consumer is currently processing.
+/// - The truth that the producer published a new back buffer since the last
+///   consumer update.
+///
+/// See the `BACK_XYZ` constants for layout details.
+///
+/// This amount of information fits in a [`u8`]. However `BackInfo` is a
+/// [`usize`] because it is frequently updated using atomic operations and we
+/// would rather not take bets about how the efficiency of sub-word atomic
+/// operations on the target CPU.
+type BackInfo = usize;
 //
-// These types are used to index into triple buffers. In addition, the
-// BackBufferInfo type is actually a bitfield, whose third bit (numerical
-// value: 4) is set to 1 to indicate that the producer published an update into
-// the back-buffer, and reset to 0 when the consumer fetches the update.
+/// Atomic version of [`BackInfo`]
+type AtomicBackInfo = AtomicUsize;
 //
-type BufferIndex = u8;
-type BackBufferInfo = BufferIndex;
+/// Low-order [`BackInfo`] bits track the index of the back buffer
+const BACK_IDX_MASK: BackInfo = 0b0_11;
 //
-type AtomicBackBufferInfo = AtomicU8;
-const BACK_INDEX_MASK: u8 = 0b11; // Mask used to extract back-buffer index
-const BACK_DIRTY_BIT: u8 = 0b100; // Bit set by producer to signal updates
+/// The next bit tracks the truth that the producer has updated the back buffer
+/// since the last consumer update..
+const BACK_UPDATED_BIT: BackInfo = 0b1_00;
+//
+/// Initial [`Status`]: back buffer is #0, no update
+const BACK_INFO_INIT: BackInfo = 0b0_00;
+
+/// Either a valid buffer index from 0 to 2, or a [`NO_BUFFER_INDEX`] sentinel
+/// value denoting an absence of buffer index
+type MaybeBufferIdx = BufferIdx;
+//
+/// Sentinel value stored in a [`MaybeBufferIdx`] to indicate an absence of
+/// buffer index.
+const NO_BUFFER_IDX: MaybeBufferIdx = 0b11;
+
+/// A pair of [`MaybeBufferIdx`] packed into a single [`AtomicU8`], representing
+/// the next input and output buffer indices if any.
+///
+/// Using the input index for the sake of illustration...
+///
+/// - If no [`Input`] currently exists, then the next input index is the buffer
+///   index that a newly created [`Input`] should use
+///   * At initialization time this is a hardcoded initial input index
+///   * If an [`Input`] existed before, but has been destroyed, then this
+///     is a backup of the previous [`Input::input_idx`], to be reused
+///     next time an `Input` is created.
+/// - As long an [`Input`] exists, this is [`NO_BUFFER_IDX`], because the
+///   input index is tracked by the [`Input`] not the shared state.
+///
+/// The output index works identically, but with [`Output`] taking the place
+/// of [`Input`], and the hardcoded initial output index is different.
+///
+/// This is a [`u8`] even though sub-word atomics may be inefficient, because
+/// this data is only accessed when an [`Input`] or [`Output`] is built or
+/// destroyed, and that operation should be rare in well-behaved client code.
+///
+/// For the same reason, this data can be packed in the same cache line as the
+/// [`BackInfo`], as penalties related to this false sharing should not be paid
+/// often. Packing two indices into a single [`u8`] is fair game too, and this
+/// allows creating an [`Input`]/[`Output`] pair using a single atomic
+/// transaction as a small efficiency bonus.
+type NextIndices = u8;
+//
+/// Atomic version of [`NextIndices`]
+type AtomicNextIndices = AtomicU8;
+//
+/// Mask covering the next input index
+const NEXT_INPUT_MASK: NextIndices = 0b00_11;
+//
+/// Mask covering the next output index
+const NEXT_OUTPUT_MASK: NextIndices = 0b11_00;
+//
+/// Initial [`Status`]: input buffer is #1, output buffer is #2
+const NEXT_INDICES_INIT: NextIndices = 0b10_01;
+//
+/// Absence of next input and output index
+///
+/// Per the above definitions, this is the expected value of the next indices of
+/// a [`TripleBuffer`] when it has both [`Input`] and an [`Output`].
+const NEXT_NOTHING: NextIndices = (NO_BUFFER_IDX << NEXT_INPUT_MASK.trailing_zeros())
+    | (NO_BUFFER_IDX << NEXT_OUTPUT_MASK.trailing_zeros());
 
 /// Unit tests
 #[cfg(test)]
 mod tests {
-    use super::{BufferIndex, SharedState, TripleBuffer, BACK_DIRTY_BIT, BACK_INDEX_MASK};
+    use super::{BufferIdx, TripleBuffer, STATUS_BACK_INDEX_MASK, STATUS_DIRTY_BIT};
     use std::{fmt::Debug, ops::Deref, sync::atomic::Ordering, thread, time::Duration};
     use testbench::race_cell::{RaceCell, Racey};
 
@@ -666,7 +990,7 @@ mod tests {
         assert!(unsafe {
             !dummy_state.eq(&SharedState::<u16>::new(
                 |i| [111, 222, 333][i],
-                BACK_DIRTY_BIT & 0b10,
+                STATUS_DIRTY_BIT & 0b10,
             ))
         });
         assert!(unsafe { !dummy_state.eq(&SharedState::<u16>::new(|i| [111, 222, 333][i], 0b01)) });
@@ -707,7 +1031,7 @@ mod tests {
     #[test]
     fn clone_shared() {
         // Let's create some dummy shared state
-        let dummy_state = SharedState::<u8>::new(|i| [123, 231, 132][i], BACK_DIRTY_BIT & 0b01);
+        let dummy_state = SharedState::<u8>::new(|i| [123, 231, 132][i], STATUS_DIRTY_BIT & 0b01);
 
         // Now, try to clone it
         let dummy_state_copy = unsafe { dummy_state.clone() };
@@ -716,7 +1040,7 @@ mod tests {
         assert!(unsafe {
             dummy_state.eq(&SharedState::<u8>::new(
                 |i| [123, 231, 132][i],
-                BACK_DIRTY_BIT & 0b01,
+                STATUS_DIRTY_BIT & 0b01,
             ))
         });
 
@@ -738,8 +1062,8 @@ mod tests {
         }
         buf.input
             .shared
-            .back_info
-            .store(BACK_DIRTY_BIT & 0b01, Ordering::Relaxed);
+            .status_bits
+            .store(STATUS_DIRTY_BIT & 0b01, Ordering::Relaxed);
         buf.input.input_idx = 0b10;
         buf.output.output_idx = 0b00;
 
@@ -764,8 +1088,8 @@ mod tests {
             assert_eq!(*buf.input.shared.buffers[2].get(), 5.6);
         }
         assert_eq!(
-            buf.input.shared.back_info.load(Ordering::Relaxed),
-            BACK_DIRTY_BIT & 0b01
+            buf.input.shared.status_bits.load(Ordering::Relaxed),
+            STATUS_DIRTY_BIT & 0b01
         );
         assert_eq!(buf.input.input_idx, 0b10);
         assert_eq!(buf.output.output_idx, 0b00);
@@ -779,8 +1103,8 @@ mod tests {
         let old_buf = buf.clone();
         let old_input_idx = old_buf.input.input_idx;
         let old_shared = &old_buf.input.shared;
-        let old_back_info = old_shared.back_info.load(Ordering::Relaxed);
-        let old_back_idx = old_back_info & BACK_INDEX_MASK;
+        let old_back_info = old_shared.status_bits.load(Ordering::Relaxed);
+        let old_back_idx = old_back_info & STATUS_BACK_INDEX_MASK;
         let old_output_idx = old_buf.output.output_idx;
 
         // Check that updating from a clean state works
@@ -795,8 +1119,8 @@ mod tests {
         expected_buf
             .input
             .shared
-            .back_info
-            .store(old_input_idx | BACK_DIRTY_BIT, Ordering::Relaxed);
+            .status_bits
+            .store(old_input_idx | STATUS_DIRTY_BIT, Ordering::Relaxed);
         assert_eq!(buf, expected_buf);
         check_buf_state(&mut buf, true);
 
@@ -807,8 +1131,8 @@ mod tests {
         expected_buf
             .input
             .shared
-            .back_info
-            .store(old_back_idx | BACK_DIRTY_BIT, Ordering::Relaxed);
+            .status_bits
+            .store(old_back_idx | STATUS_DIRTY_BIT, Ordering::Relaxed);
         assert_eq!(buf, expected_buf);
         check_buf_state(&mut buf, true);
 
@@ -818,7 +1142,7 @@ mod tests {
         expected_buf
             .output
             .shared
-            .back_info
+            .status_bits
             .store(old_output_idx, Ordering::Relaxed);
         assert_eq!(buf, expected_buf);
         check_buf_state(&mut buf, false);
@@ -837,8 +1161,8 @@ mod tests {
             buffer.push(2);
 
             // not yet published
-            let back_info = buffer.reference.shared.back_info.load(Ordering::Relaxed);
-            let back_buffer_dirty = back_info & BACK_DIRTY_BIT != 0;
+            let back_info = buffer.reference.shared.status_bits.load(Ordering::Relaxed);
+            let back_buffer_dirty = back_info & STATUS_DIRTY_BIT != 0;
             assert!(!back_buffer_dirty);
         }
         check_buf_state(&mut buf, true); // after publish, before read
@@ -1161,7 +1485,7 @@ mod tests {
 
     /// Range check for triple buffer indexes
     #[allow(unused_comparisons)]
-    fn index_in_range(idx: BufferIndex) -> bool {
+    fn index_in_range(idx: BufferIdx) -> bool {
         (0..=2).contains(&idx)
     }
 
@@ -1182,9 +1506,9 @@ mod tests {
         assert_eq!(as_ptr(&buf.input.shared), as_ptr(&buf.output.shared));
 
         // Access the shared state and decode back-buffer information
-        let back_info = buf.input.shared.back_info.load(Ordering::Relaxed);
-        let back_idx = back_info & BACK_INDEX_MASK;
-        let back_buffer_dirty = back_info & BACK_DIRTY_BIT != 0;
+        let back_info = buf.input.shared.status_bits.load(Ordering::Relaxed);
+        let back_idx = back_info & STATUS_BACK_INDEX_MASK;
+        let back_buffer_dirty = back_info & STATUS_DIRTY_BIT != 0;
 
         // Input-/output-/back-buffer indexes must be in range
         assert!(index_in_range(buf.input.input_idx));